Adding parsing of rill iso extensions
