+++ runtime/pkg/duration/duration_test.go
@@ -9,14 +9,14 @@ import (
 func TestParseISO8601(t *testing.T) {
 	tests := []struct {
 		from     string
-		expected Duration
+		expected StandardDuration
 		err      bool
 	}{
-		{from: "P2W", expected: Duration{Week: 2}},
-		{from: "P1Y2WT5M", expected: Duration{Year: 1, Week: 2, Minute: 5}},
+		{from: "P2W", expected: StandardDuration{Week: 2}},
+		{from: "P1Y2WT5M", expected: StandardDuration{Year: 1, Week: 2, Minute: 5}},
 		{from: "P1X", err: true},
-		{from: "inf", expected: Duration{Inf: true}},
-		{from: "Inf", expected: Duration{Inf: true}},
+		{from: "inf", expected: StandardDuration{Inf: true}},
+		{from: "Inf", expected: StandardDuration{Inf: true}},
 		{from: "infinity", err: true},
 	}
 	for _, tt := range tests {
